Implemented funcition to compute averages of each category and sort them in RecCenters.js
computeAverages() reads in data from backend, computes the average rating of the specified category, and returns it to clickSort(). clickSort() then sorts the averages from high -> low and returns an alert message for now (still need to figure out how to update the frontend)
Also, I modified the review button name for Dorms and Dining Halls.